requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
mongoengine
python-multipart
python-jose[cryptography]
pycryptodome
passlib
bcrypt


